moved messages to hashers internal
